+++ features/ajna/controls/AjnaFormWrapper.tsx
@@ -4,7 +4,7 @@ import { ethers } from 'ethers'
 import { AjnaBorrowFormState } from 'features/ajna/borrow/state/ajnaBorrowFormReducto'
 import { useAjnaTxHandler } from 'features/ajna/borrow/useAjnaTxHandler'
 import { AjnaBorrowAction, AjnaSidebarStep } from 'features/ajna/common/types'
-import { useAjnaGeneralContext } from 'features/ajna/contexts/AjnaProductContext'
+import { useAjnaGeneralContext } from 'features/ajna/contexts/AjnaGeneralContext'
 import { useAccount } from 'helpers/useAccount'
 import { useFlowState } from 'helpers/useFlowState'
 import { zero } from 'helpers/zero'
