+++ features/ajna/controls/AjnaProductController.tsx
@@ -7,7 +7,7 @@ import { steps } from 'features/ajna/common/consts'
 import { getAjnaHeadlineProps } from 'features/ajna/common/helpers'
 import { AjnaWrapper } from 'features/ajna/common/layout'
 import { AjnaFlow, AjnaProduct } from 'features/ajna/common/types'
-import { AjnaGeneralContextProvider } from 'features/ajna/contexts/AjnaProductContext'
+import { AjnaGeneralContextProvider } from 'features/ajna/contexts/AjnaGeneralContext'
 import { AjnaEarnView } from 'features/ajna/earn/views/AjnaEarnView'
 import { WithTermsOfService } from 'features/termsOfService/TermsOfService'
 import { WithWalletAssociatedRisk } from 'features/walletAssociatedRisk/WalletAssociatedRisk'
