+++ features/ajna/earn/views/AjnaEarnView.tsx
@@ -3,7 +3,7 @@ import { TabBar } from 'components/TabBar'
 import { VaultHeadline } from 'components/vault/VaultHeadline'
 import { useAjnaBorrowContext } from 'features/ajna/borrow/contexts/AjnaBorrowContext'
 import { getAjnaHeadlineProps } from 'features/ajna/common/helpers'
-import { useAjnaGeneralContext } from 'features/ajna/contexts/AjnaProductContext'
+import { useAjnaGeneralContext } from 'features/ajna/contexts/AjnaGeneralContext'
 import { AjnaEarnOverviewWrapper } from 'features/ajna/earn/overview/AjnaEarnOverviewWrapper'
 import { AjnaEarnFormWrapper } from 'features/ajna/earn/sidebars/AjnaEarnFormWrapper'
 import { VaultOwnershipBanner } from 'features/notices/VaultsNoticesView'
